requirements.txt
Django==4.1.5

psycopg2-binary==2.9.5

django-bootstrap-form==3.4

django-bootstrap-datepicker-plus==5.0.3

uwsgi == 2.0.21

invoke == 1.7.3

boto3==1.26.50
pexpect~=4.8.0
click~=8.1.3